+++ changelog.md
@@ -19,6 +19,7 @@ Version 0.0.3 (August ?, 2018)
 ### test
 
  - **http:** add unitary tests
+ - **link:** add unitary tests
  - **navigation:** add unitary tests
 
 Version 0.0.2 (August 10, 2018)